dev: refresh InfiniteLongLM tree from 2026-05-30 zip
Headline change: new per-q-head landmark-key path lays groundwork for
high-GQA-ratio configs (e.g. h_q=16, h_kv=4 -> lmk_k stays 16 heads,
not collapsed to 4). New ops files added:
- ops/topk_head.py: max-pooling topk supporting both shared
  [B,S,h_kv,D] and per-q-head [B,S,h_kv*G,D] lmk shapes.
- ops/hsa_fwd_bwd_single_tilelang.py: head-wise HSA reference with
  per-q-head weights (B, q_blocks, HQ, K), no shared-across-G assumption.

Real code changes in models/FlashHSA/lhsa_layer.py:
- New `shared_q_c` config: chunk_pool mu_q comes from a learnable
  nn.Parameter(h_q, head_dim) shared across all chunks instead of the
  per-chunk last lmk_q (default False).
- New `enable_chunk_pooling` config: replaces lmk_k surrogate with
  chunk-inner mean-pool over K (default False). Cross-product with
  `compact_lmk_k` gives 4 lmk_k modes (last-token / mean-pool /
  softmax-weighted-pool).
- chunk_attn_pool `mu_g` uses reshape (not view) so non-contiguous
  mu_q from shared_q_c's `q_c.expand(...)` works.
- `use_gumbel` arg removed from topk_func call.

ops/topk_head_softmax.py drops use_gumbel entirely and adds varlen
mode (cu_seq_lens, B=1 packed, dynamic seq/s_len) for SFT sequence
packing.

For our canonical 1-layer align config (hsa_heads=8, hsa_qk_ratio=1,
G=1), the new per-q-head path is a no-op and forward output is
numerically identical to the 2026-05-22 snapshot. Verified: compare.py
all 9 cases PASS, 64-req churn stress KL prefill mean=0.003657
max=0.004187, decode mean=0.004370 max=0.006380, zero outliers.

Excludes dev/InfiniteLongLM.{zip,prev}/, "eval_sglang copy.sh", HF
token in download_dolma3*.sh redacted to ${HF_TOKEN}.

+++ dev/InfiniteLongLM/models/FlashHSA/modeling_qwen_lhsa.py
@@ -535,7 +535,7 @@ class Qwen3RotaryEmbedding(nn.Module):
             cos = emb.cos() * self.attention_scaling
             sin = emb.sin() * self.attention_scaling
 
-        return cos, sin
+        return cos.to(dtype=x.dtype), sin.to(dtype=x.dtype)
 
 
 QWEN3_START_DOCSTRING = r"""
